// R1.2_Documents/51_製造/21_参考資料/HTML_FR-CFM-DFM_起動手順/HTML_FR-CFM-DFM/source/src/app/types/flm/car.ts
export interface CarIdSearchParams {
  common: {
    car_identification: {
      car_ids: string[];
    };
  };
}
// ...
export interface TimeDifferenceSettingParams {
  cars: {
    update_datetime: string;
    time_difference: string;
    car_id: string;
  }[];
}